various

+++ workflows/analysis/hexmaps/__main__.py
@@ -1,5 +1,5 @@
 import os
-from utils import Log
+from utils import Log, File, Hash, JSONFile
 import shutil
 import random
 
@@ -11,12 +11,15 @@ def clean_and_copy():
     for dirpath, dirnames, filenames in os.walk(dir_root):
         for filename in filenames:
             file_path = os.path.join(dirpath, filename)
-
             if filename.endswith(".py"):
                 continue
+            if filename.endswith(".hash.json"):
+                continue
             if filename.startswith("hexbin-") and filename.endswith(".png"):
-                shutil.move(file_path, os.path.join("images", filename))
+                image_path = os.path.join("images", filename)
+                shutil.move(file_path, image_path)
                 log.info(f'Moved "{file_path}" to images')
+                os.startfile(image_path)
                 continue
             os.remove(file_path)
             log.warning(f'Removed "{file_path}"')
@@ -26,6 +29,33 @@ def clean_and_copy():
             if dirname == "__pycache__":
                 shutil.rmtree(dir_path)
                 log.warning(f'Removed "{dir_path}"')
+
+
+def run_single(file_path):
+    content = File(file_path).read()
+    VERSION = "v2"
+    md5 = Hash.md5(content + VERSION)
+
+    hash_json_file_path = file_path[:-3] + ".hash.json"
+    hash_json_file = JSONFile(hash_json_file_path)
+    current_md5 = None
+    if os.path.exists(hash_json_file_path):
+        data = hash_json_file.read()
+        current_md5 = data["md5"]
+
+    if current_md5 == md5:
+        log.info(f"Not running {file_path}")
+        return
+
+    cmd = f"python {file_path}"
+    print("-" * 60)
+    log.info(f"{cmd}")
+    os.system(cmd)
+    print("-" * 60)
+
+    hash_json_file.write(dict(md5=md5))
+
+    clean_and_copy()
 
 
 def run():
@@ -41,17 +71,13 @@ def run():
     random.shuffle(exe_path_list)
 
     for file_path in exe_path_list:
-        cmd = f"python {file_path}"
-        print("-" * 60)
-        log.info(f"{cmd}")
-        os.system(cmd)
-        print("-" * 60)
-        clean_and_copy()
+        run_single(file_path)
 
 
 def main():
 
     run()
+    clean_and_copy()
 
 
 if __name__ == "__main__":
